bugfixes in compute_types, add residual_flux2 in elliptic
compute_types: brace-init' lists of the form
```std::array<double, 3> arr = {.1};```
only initialize the first element, the rest are value-init', in this
case zeroed.

+++ experiments/read_domain.cxx
@@ -11,7 +11,7 @@ int main(int argc, char** argv) {
     return 1;
   }
 
-  auto info = read_domain_geobin<1, 3,  std::vector, Point<3, double>, uint64_t, uint64_t, uint64_t>(argv[1]);
+  auto info = read_domain_geobin<1, 3,  std::vector, Point<3, double>, uint32_t, uint32_t, uint32_t>(argv[1]);
 
   std::println("{} {} {}", info.points.size(), info.hyNodes_hyEdge.size(), info.hyEdge_properties.size());
 }
